Module9/src/requirements.txt
functions-framework==3.*
blinker==1.6.2

cachetools==5.3.1

certifi==2023.5.7

charset-normalizer==3.1.0

click==8.1.3

colorama==0.4.6

Flask-Cors==4.0.0

enum34==1.1.10

Flask==2.3.2

google-api-core==2.11.0

google-api-python-client==2.88.0

google-auth==2.19.1

google-auth-httplib2==0.1.0

google-cloud==0.34.0

google-cloud-core==2.3.2

google-cloud-firestore==2.11.1

google-cloud-language==2.10.0

googleapis-common-protos==1.59.0

grpcio==1.54.2

grpcio-status==1.54.2

httplib2==0.22.0

idna==3.4

itsdangerous==2.1.2

Jinja2==3.1.2

MarkupSafe==2.1.3

proto-plus==1.22.2

protobuf==4.23.2

pyasn1==0.5.0

pyasn1-modules==0.3.0

pyparsing==3.0.9

requests==2.31.0

rsa==4.9

six==1.16.0

uritemplate==4.1.1

urllib3==1.26.16

Werkzeug==2.3.5

boto3==1.16.0